requirements.txt
numpy==1.21.6
pandas==1.3.5
opencv-python==4.8.0.76
pygame==2.5.2
moviepy==1.0.3
absl-py==2.1.0
matplotlib==3.5.3
#gym==0.19.0